+++ Template/dashboard/addshortcuts_on_dashboard_sidebar.php
@@ -1,0 +1,3 @@
+<li <?= $this->app->checkMenuSelection('AddShortcutsController', 'viewShortcutPresetSelectModal') ?>>
+    <a href="/addshortcuts/shortcutPresetSelectModal" class="js-modal-small">Shortcuts</a>
+</li>